// src/lib/types.ts
export type Product = {
  id: string;
  name: string;
  category: string;
  location: string;
  quantity: number;
  minQuantity: number;
  idealQuantity: number;
  unit: string;
  brand?: string;
  barcode?: string;
  image?: string;
  favorite: boolean;
  inCart?: boolean;
  updatedAt: string;
};

export type Category = { id: string; name: string; emoji?: string };
export type Location = { id: string; name: string };